requirements.txt
Twisted
smpp.pdu
chardet==2.2.1